refactor(wasm): extract import-scan pass to _scan.py
_scan.py now owns the CFG-module walk that classifies each command
against the runtime-import tables and returns the set of runtime
functions the emitter needs. Removes ~630 lines of scan dispatch
from __init__.py, leaving the emitter class itself much easier to
navigate.

_scan.py imports:
- ValType / WasmOp indirectly through consumers
- the dispatch tables from _imports.py
- _parse_array_ref / _has_embedded_subst_scan from _parsing.py
- IR node types from ../../ir
- ExprAst nodes from ../../expr_ast

Public API unchanged: __init__.py re-exports _scan_needed_imports,
_scan_text_for_cmd_subst, and _scan_expr_body_imports for callers
that still import them by name.

__init__.py is now ~5956 lines (from 7726). Five submodules hold
~2060 lines of extracted code. make prep-pr exits 0: 9985 tests,
lint, and typecheck all green.
